Port r82630 (bwilson) from llvm-gcc: Improve handling of bitfields in unions by sharing more of the code that handles bitfields in structs.  Radar 7226434.
The immediate problem is that a union containing only a bitfield may cause
an assertion failure.  GCC sets the TREE_TYPE of a field to the smallest type
that contains the bitfield, and llvm had just been using that type (except
for a special case that Dale added in svn r47655 to handle bool bitfields on
PowerPC).  For a 24-bit field on ARM, TREE_TYPE was 32 bits but the size of
the union was only 24 bits, and this caused an assertion failure.  (We didn't
see this on X86 because the minimum field alignment for integer types is 32
bits, so the union size was rounded up to 32 bits.)

The struct conversion code already handles all of this, so I merged the
ConvertUNION method into ConvertRecord to reuse all the existing logic.
I also changed ConvertUnionCONSTRUCTOR to reuse a lot of the code from
ConvertRecordCONSTRUCTOR.



git-svn-id: https://llvm.org/svn/llvm-project/dragonegg/trunk@85101 91177308-0d34-0410-b5e6-96231b3b80d8