Port commit 100721 (stoklund) from llvm-gcc: Create larger struct fields for consecutive bitfields. (Again).
When creating an LLVM struct type for a struct with bitfields, try to allocate
integer fields that are as large as possible, taking the following bitfields
into account.

For example, this struct:

  struct S {
    int a, b;
    void *c;
    unsigned d : 8;
    unsigned e : 8;
  };

used to be:

  %struct.S = type { i32, i32, i8*, i8, i8 },

but now it becomes:

  %struct.S = type { i32, i32, i8*, i16 }

This makes it easier for the code generator to shuffle register sized bitfield
groups around in fewer registers. It also means that it is harder for SROA to
isolate bit fields as scalars.

The larger struct fields are only produced when they would be properly aligned
and smaller than the natural machine word size.

This strategy is similar to what clang will be doing. We will see how it works
out.



git-svn-id: https://llvm.org/svn/llvm-project/dragonegg/trunk@103967 91177308-0d34-0410-b5e6-96231b3b80d8