Backport the fix for PR13561 (r168868) to the 3.2 branch.  Original commit message:
Fix PR13561, a crash when the GCC vectorizer is turned on.  What happens here is
that a vector with type <2 x i64> is being initialized with <i1 1, i1 1>, a type
mismatch.  This is coming directly from GCC, which is pretty relaxed about types
in general, expecting implicit casts somewhat like in C.  Fix this by turning
GCC's implicit type cast into an explicit LLVM one.


git-svn-id: https://llvm.org/svn/llvm-project/dragonegg/branches/release_32@168870 91177308-0d34-0410-b5e6-96231b3b80d8